ocr_orderbook_server/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
flask>=3.0.0
pyyaml>=6.0
yt-dlp>=2024.1.0
ffmpeg-python>=0.2.0
